+++ src/lib/index.ts
@@ -1,7 +1,10 @@
 // Images & icons
 export {default as icon} from './images/icon.svg';
-export {default as darkLogo} from './images/dark-logo.svg';
-export {default as lightLogo} from './images/light-logo.svg';
+export {default as logo} from './images/logo.svg';
+export {default as logoBeside} from './images/logo-beside.svg';
+export {default as github} from './images/github.svg';
+export {default as session} from './images/session.svg';
+export {default as tor} from './images/tor.svg';
 
 // Svelte Components
 export {default as Header} from './components/Header.svelte';