+++ src/routes/+page.svelte
@@ -1,22 +1,23 @@
 <script lang="ts">
-  import {darkLogo} from '$lib';
+  import {logo} from '$lib';
+  import {getContext} from 'svelte';
+  import type {Writable} from 'svelte/store';
+
+  const authenticated: Writable<boolean> = getContext('auth');
 </script>
 
-<img src={darkLogo} alt="App" width="512" height="512" />
-<p>click that button to start developing the app!</p>
-<button>Begin 😃</button>
-
+<div>
+  <img src={logo} alt="App" width="512" height="512" />
+  <p>Click that button to start your amazing journey!</p>
+  <button on:click={() => ($authenticated = !$authenticated)}>Begin 😃</button>
+</div>
 
 <style lang="postcss">
-  img {
-    @apply mt-52;
-  }
-
-  p {
-    @apply mt-8 text-neutral-300 text-4xl;
+  div {
+    @apply mt-24 flex flex-col items-center justify-center gap-8;
   }
 
   button {
-    @apply mt-8 text-4xl text-neutral-300 bg-primary-700 p-6 rounded-md hover:bg-primary-900 shadow-xl shadow-primary-900 transition-all hover:shadow-primary-950;
+    @apply p-6 text-3xl;
   }
 </style>